refactor to more consistently use dependency injection, facilitate unit tests
