+++ src/Queue/IBlockingPriorityQueue.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleL7Proxy.Queue;
+
+public interface IBlockingPriorityQueue<T>
+{
+  int MaxQueueLength { get; }
+  void StartSignaler(CancellationToken cancellationToken);
+  void Stop();
+  int Count { get; }
+  bool Enqueue(T item, int priority, DateTime timestamp);
+  bool Requeue(T item, int priority, DateTime timestamp);
+  Task<T> Dequeue(CancellationToken cancellationToken, string id);
+}
