+++ src/Queue/QueueServiceCollectionExtensions.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleL7Proxy.Queue
+{
+  internal class QueueServiceCollectionExtensions
+  {
+  }
+}
